WIP: Subscription tests

+++ src/Plan.php
@@ -17,21 +17,68 @@ class Plan extends Model
      */
     protected $guarded = [];
 
-    public static function createAsConektaPlan($plan_id, $attributes)
+    public static function createAsConektaPlan($name, $attributes)
     {
-        $attributes = array_merge(['id' => $plan_id], $attributes);
+        $trial_ends_at = null;
+        if ($attributes['trial_ends_at'] && $attributes['trial_ends_at'] instanceof Carbon) {
+            $trial_ends_at = $attributes['trial_ends_at'];
+            $attributes['trial_period_days'] = Carbon::now()->diffInDays($attributes['trial_ends_at']);
+            unset($attributes['trial_ends_at']);
+        }
+
+        $plan_id = self::generatePlanId($attributes);
+
+        $attributes = array_merge([
+            'id' => $plan_id,
+            'name' => $name,
+            'amount' => 1000,
+            'currency' => 'MXN',
+            'interval' => 'month',
+            'frequency' => 1,
+            'trial_period_days' => null,
+            'expiry_count' => null,
+        ], $attributes);
 
         try {
             $plan = ConektaPlan::find($plan_id);
-            $plan->delete();
         } catch (\Exception $e) {
+            $plan = null;
+        }
+        
+        if ($plan) {
+            $plan->delete();
         }
 
         $plan = ConektaPlan::create($attributes);
-        
-        self::create(['conekta_id' => $plan->id]);
+
+        self::updateOrCreate(['conekta_id' => $plan->id], [
+            'conekta_id' => $plan->id,
+            'trial_ends_at' => $trial_ends_at,
+        ]);
 
         return $plan;
+    }
+
+    private static function generatePlanId($attributes)
+    {
+        $only = [   
+            'currency',
+            'amount',
+            'interval',
+            'frequency',
+            'trial_period_days',
+            'trial_ends_at',
+            'expiry_count',
+        ];
+
+        $attributes = array_filter($attributes, function ($key) use ($only) {
+            return in_array($key, $only);
+        }, ARRAY_FILTER_USE_KEY);
+
+        // Sort the array
+        $attributes = array_replace(array_flip($only), $attributes);
+
+        return strtolower(implode('-', $attributes));
     }
 
     /**
